# specs/001-piano-teacher/contracts/api.yaml
openapi: 3.0.3
info:
  title: Piano Teacher Management System API
  description: RESTful API for managing piano students, lessons, schedules, and homework
  version: 1.0.0
  contact:
    name: Piano Teacher System
    email: support@pianoteacher.com
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
  - url: https://api.pianoteacher.com/v1
    description: Production server
paths:
...
  /lessons/{id}:
    get:
      summary: Get lesson by ID
      description: Retrieve detailed information about a specific lesson
      tags:
        - Lessons
      parameters:
        - name: id
          in: path
          required: true
          description: Lesson ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Lesson details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lesson'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'

    put:
      summary: Update lesson
      description: Update lesson information
      tags:
        - Lessons
      parameters:
        - name: id
          in: path
          required: true
          description: Lesson ID
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateLessonRequest'
      responses:
        '200':
          description: Lesson updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lesson'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
    Student:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          maxLength: 100
          example: "张小明"
        age:
          type: integer
          minimum: 5
          maximum: 100
          example: 12
        grade:
          type: string
          maxLength: 50
          example: "六年级"
        email:
          type: string
          format: email
          example: "student@example.com"
        phone:
          type: string
          maxLength: 20
          example: "+86-138-0013-8000"
        notes:
          type: string
          maxLength: 1000
          example: "对音乐有浓厚兴趣，练习认真"
        pricing:
          $ref: '#/components/schemas/Pricing'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Pricing:
      type: object
      properties:
        id:
          type: integer
          format: int64
        hourlyRate:
          type: number
          format: decimal
          example: 150.00
        lessonDuration:
          type: number
          format: decimal
          example: 1.0
        paymentTerms:
          type: string
          maxLength: 500
          example: "月结"
...
    Lesson:
      type: object
      properties:
        id:
          type: integer
          format: int64
        startTime:
          type: string
          format: date-time
          example: "2025-01-30T14:00:00"
        endTime:
          type: string
          format: date-time
          example: "2025-01-30T15:00:00"
        topics:
          type: string
          maxLength: 1000
          example: "音阶练习，C大调"
        techniques:
          type: string
          maxLength: 1000
          example: "手指独立性训练"
        materials:
          type: string
          maxLength: 500
          example: "拜厄练习曲第20条"
        notes:
          type: string
          maxLength: 2000
          example: "学生进步明显，节奏感增强"
        status:
          $ref: '#/components/schemas/LessonStatus'
        student:
          $ref: '#/components/schemas/Student'
        homework:
          type: array
          items:
            $ref: '#/components/schemas/Homework'
        performance:
          $ref: '#/components/schemas/Performance'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    UpdateLessonRequest:
      type: object
      properties:
        startTime:
          type: string
          format: date-time
          example: "2025-01-30T14:00:00"
        endTime:
          type: string
          format: date-time
          example: "2025-01-30T15:00:00"
        topics:
          type: string
          maxLength: 1000
          example: "音阶练习，C大调"
        techniques:
          type: string
          maxLength: 1000
          example: "手指独立性训练"
        materials:
          type: string
          maxLength: 500
          example: "拜厄练习曲第20条"
        notes:
          type: string
          maxLength: 2000
          example: "学生进步明显，节奏感增强"
        status:
          $ref: '#/components/schemas/LessonStatus'
    LessonStatus:
      type: string
      enum:
        - SCHEDULED
        - COMPLETED
        - CANCELLED
        - RESCHEDULED
    Homework:
      type: object
      properties:
        id:
          type: integer
          format: int64
        description:
          type: string
          maxLength: 2000
          example: "练习拜厄第20-25条，注意节奏准确性"
        dueDate:
          type: string
          format: date
          example: "2025-02-06"
        status:
          $ref: '#/components/schemas/HomeworkStatus'
        completionNotes:
          type: string
          maxLength: 1000
          example: "完成度良好，第23条需要加强练习"
        lesson:
          $ref: '#/components/schemas/Lesson'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    HomeworkStatus:
      type: string
      enum:
        - ASSIGNED
        - IN_PROGRESS
        - COMPLETED
        - OVERDUE
    Performance:
      type: object
      properties:
        id:
          type: integer
          format: int64
        rating:
          type: integer
          minimum: 1
          maximum: 10
          example: 8
        notes:
          type: string
          maxLength: 2000
          example: "音阶演奏流畅，和弦转换需要加强"
        improvements:
          type: string
          maxLength: 1000
          example: "加强左手力度练习"
        achievements:
          type: string
          maxLength: 1000
          example: "掌握了C大调音阶"
        evaluationDate:
          type: string
          format: date-time
          example: "2025-01-30T15:00:00"
        lesson:
          $ref: '#/components/schemas/Lesson'
...
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string
  responses:
    BadRequest:
      description: Bad request - invalid input data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...